requirements_antigo.txt
APScheduler==3.6.3
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
boto3==1.26.94
botocore==1.29.94
cachetools==4.2.2
certifi==2022.9.24
charset-normalizer==2.0.12
crispy-bootstrap5==0.6
Deprecated==1.2.13
Django==3.2.13
django-auto-prefetch==1.3.0
django-colorfield==0.8.0
django-crispy-forms==1.13.0
django-currentuser==0.5.3
django-extensions==3.1.5
django-filter==23.2
django-multiselectfield==0.1.12
django-ratelimit==4.0.0
django-sse==0.4.1
django-storages==1.13.2
exceptiongroup==1.0.4
gunicorn==20.1.0
h11==0.14.0
idna==3.4
jmespath==1.0.1
outcome==1.2.0
packaging==21.3
Pillow==9.4.0
psutil==5.9.1
psycopg2==2.9.3
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-telegram-bot==13.12
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
redis==4.3.3
requests==2.27.1
s3transfer==0.6.0
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
sse==1.2
tornado==6.2
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.6
tzlocal==4.2
undetected-chromedriver==3.1.5.post4
urllib3==1.26.13
websocket-client==1.3.3
websockets==10.4
whitenoise==6.2.0
wrapt==1.14.1
wsproto==1.2.0